pyproject.toml
[project]
name = "network-traffic-monitoring"
version = "0.1.0"
description = ""
authors = [
    {name = "Mouhamadou Bamba DIOP [ext_diop004907 DRPS/PEX/SDN/SSD]",email = "MouhamadouBamba.DIOP@orange-sonatel.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy (>=2.4.0,<3.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "elasticsearch (>=9.2.1,<10.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "network_traffic_monitoring", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
